% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_information.R
\name{add_info}
\alias{add_info}
\title{Add latlon coordinates or other info to an EnvLogger tibble}
\usage{
add_info(env, info)
}
\arguments{
\item{env}{The output of a call to \code{\link[=read_env]{read_env()}} (either directly, of the \verb{$report} element). Likely, a call to \code{\link[=parse_id]{parse_id()}} will also be required to isolate the location name component from the logger ids.}

\item{info}{A data.frame (or tibble) providing location names, latlon coordinates, and any additional information to be joined to \code{env}. The column coding for location names must be strictly named \code{site}, and must be present. Columns coding for latitude and longitude, if included, must be strictly named \code{lat} and \code{lon}, not \code{long}, \code{longitude} or \verb{x,y}. All additional columns are retained and joined (any column name matching to colnames in \code{env} will receive the suffix \code{x_}).}
}
\value{
The same structure provided as input (\code{env}), augmented with new columns resulting from the joining of the information provided.
}
\description{
Joins latlon or other relevant information to an EnvLogger tibble resulting from a call to \code{\link[=read_env]{read_env()}}. Information is joined by location name, so, unless the logger ids used originally code for location name only, a call to \code{\link[=parse_id]{parse_id()}} will be required to generate a column that specifically codes for location name. Those same location names have to be used here.
}
\section{Why use \code{\link[=add_info]{add_info()}} to set latlon}{

Since EnvLogger data reports include latitude and longitude, it is reasonalbe to wonder why the \link{envlogger-package} doesn't collect latlon info from that source. In fact, the latitude and longitude provided in EnvLogger reports is there to potentially facilitate troubleshooting, in case the location of origin of a data report is uncertain due to failure to set meaningfull custom logger ids and similar issues. This is because there are many scenarious that can lead to the latlon data stored in the data report files being inaccurate or absent. For example, the smartphone used to download the data has just been turned on, a GPS fix may not yet have been available when the data was downloaded, or may be highly innacurate. The same happens if the loggers have being deployed in a location where GPS coverage is low or absent (underwater, inside of a cave, in an urban setting, etc.). On top of that, depending on the accuracy of the devices used during each download session (and whether there's only one device being used or multiple), potentially resulting in significant differences between the coordinates recorded on each file retrieved on the same day and from the same site.

With all this in mind, sourcing latlon from the data report files would be too unreliable. On top of that, users typically maintain a registry of which loggers have been deployed and where, and that is a much more trustworthy and stable source of information for site geographic coordinates. It is in this context that \code{\link[=add_info]{add_info()}} is provided, as a way to streamline the merging of the logger data imported with the curated location information.
}

\examples{
env <- read_env(
   env_example(c("ptzzy", "ptzzw")),
   new_interval = 60,
   show_progress = FALSE,
   show_warnings = FALSE)

env <- cctbon_id(env)

add_info(
   env,
   info = data.frame(
       var1 = 1:2,
       var3 = 1:2,
       site = c("ptzzy", "ptzzw"),
       var2 = 1:2,
       lat = c(36, 40),
       lon = -10,
       var4 = 1:2,
       exp = 1:2)
   )
}
