% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctbon.R
\name{cctbon_id}
\alias{cctbon_id}
\title{Decompose CCTBON ids into their elements and append to the original data}
\usage{
cctbon_id(env_data, new_sh = NULL)
}
\arguments{
\item{env_data}{a tibble generated by \code{READ_ENV(..., just_rep = TRUE)} or with equivalent format with a column named \verb{$id}, containing logger names structured according to the CCTBON scheme.}

\item{new_sh}{string, defaults to \code{NULL}; new shore name to replace any instance of shore name that doesn't comply with the standard 5-chars (nowadays always 5, but previously just 4)}
}
\value{
A tibble with \verb{$sh}, \verb{$lvl}, \verb{$exp}, \verb{$mic} and \verb{$rep}, each containing the parsed values for each element of the CCTBON id.
}
\description{
Extract country, shore, microhabitat level, exposure and replicate number from standard CCTBON ids and convert to level and exposure to factors with properly ordered factor levels.
}
\section{CCTBON IDs}{

EnvLogger custom IDs can be up to 10 characters long. To follow the CCTBON network's scheme, IDs must be structured strictly as follows:
\itemize{
\item 2 char for country code
\itemize{
\item must follow standard 2-letter country codes
\item e.g., "pt" for Portugal
}
\item 3 char for site
\itemize{
\item e.g., "ang" for Angeiras
\item any combination is accepted (including with numbers)
\item HOWEVER, the first 5 chars (country + site, i.e., the full site name) must be unique
}
\item 1 char for shore height; strictly, one of the following:
\itemize{
\item "s" for supratidal
\item "t" for top-shore
\item "m" for mid-shore
\item "l" for low-shore
\item "p" for tide-pool
}
\item 1 char for solar exposure; strictly, one of the following:
\itemize{
\item "h" for hot
\item "c" for cold
\item "w" for west
\item "e" for east
\item "i" for in "o" for out (associated with tide-pools)
}
\item 2 digits for replicate number
\itemize{
\item e.g., "01"
}
\item 1 char for sub-replicate
\itemize{
\item e.g., "a"
\item should be used sequentially in alphabetical order
}
}

For example, ID "ptminmh02a" refers to the 2nd logger deployed in the Portuguese shore of Mindelo, at the mid-shore level, with maximum exposure to the sun.
}

\examples{
paths <- env_example("ptzzy", dir = TRUE)
x <- READ_ENV(paths)
cctbon_id(x)
}
\seealso{
\code{\link[=READ_ENV]{READ_ENV()}}
}
