% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctbon.R
\name{cctbon_summarise}
\alias{cctbon_summarise}
\title{Group CCTBON data by microhabitat and summarise it}
\usage{
cctbon_summarise(
  cctbon_data,
  by_micro = FALSE,
  by_shore = FALSE,
  by_day = TRUE,
  roll_days = 0
)
}
\arguments{
\item{cctbon_data}{a tibble generated by \code{READ_ENV(..., just_rep = TRUE)}, with logger names structured according to the CCTBON scheme.}

\item{by_micro}{logical, defaults to \code{FALSE}; whether data should be grouped at the microhabitat level.}

\item{by_shore}{logical, defaults to \code{FALSE}; whether data should be grouped at the shore level (ignored if \code{by_micro = TRUE}).}

\item{by_day}{logical, defaults to \code{TRUE}; whether data should be grouped daily (applied before \code{roll_days}).}

\item{roll_days}{integer, defaults to \code{0}; number of days over which to group summarised data (rolling mean); if \code{0}, data is returned without temporal summarisation.}
}
\description{
Take CCTBON data and summarise daily and by microhabitat (i.e., discard logger by logger granularity and return min > avg > max stats for each day)
}
\examples{
paths <- env_example(c("ptzzw", "nozzz"), dir = TRUE)
x <- READ_ENV(paths)
x <- cctbon_id(x)
cctbon_summarise(x)
}
