% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{env_header_val}
\alias{env_header_val}
\title{Get the value of a given field in an Envlogger report header}
\usage{
env_header_val(header, field_pattern, force_numeric = FALSE)
}
\arguments{
\item{header}{a tibble with columns \code{field} and \code{val} and all values in lower case}

\item{field_pattern}{character string that must match exactly only one entry in the \code{field} column}

\item{force_numeric}{logical, defaults to \code{FALSE}; if \code{TRUE}, the value is stripped of non-numeric characters and converted to numeric.}
}
\value{
The value for the target field, as string or numeric (if \code{force_numeric = TRUE}).
}
\description{
Get the value of a given field in an Envlogger report header
}
\examples{
header <- tibble::tibble(field = c("field1", "field2", "field3"), val = c("a", "1", "a1"))
env_header_val(header, "field1")
env_header_val(header, "field2", force_numeric = TRUE)
env_header_val(header, "field2", force_numeric = TRUE)
}
