% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{env_ls}
\alias{env_ls}
\title{List EnvLogger files}
\usage{
env_ls(paths, avoid_pattern = NULL, list_unsupported = FALSE)
}
\arguments{
\item{paths}{A character vector of one or more paths. May include folders, which are searched recursively.}

\item{avoid_pattern}{A character vector of one or more patterns. File paths matching to \code{avoid_pattern} are dropped.}

\item{list_unsupported}{Should unsupported files be included in the output?}
}
\value{
A tibble with nine columns, including the file paths (\code{path}), two columns to indicate the type of files encountered, and six columns to facilitate quick filtering of specific file categories.

In more detail, the two columns that indicate the type of files encountered are \verb{$int} (numeric) and \verb{$chr} (character), and their values are as follows:
\itemize{
\item \code{1} and \code{report}      if an EnvLogger report or a valid report with a different format and an associated metadata file
\item \code{2} and \code{log}         if an EnvLogger logfile
\item \code{3} and \code{metadata}    if an EnvLogger metadata file
\item \code{0} and \code{unsupported} if none of the three
}

The remaining columns are logical vectors for quick filtering, namely:
\itemize{
\item \code{rep}       - path points to a valid report of any kind?
\item \code{rep_env}   - path points to an EnvLogger report?
\item \code{rep_other} - path points to a report with a different format?
\item \code{has_met}   - path points to a report with an associated metadata file?
\item \code{log}       - path points to an EnvLogger logfile?
\item \code{met}       - path points to an EnvLogger metadata file?
}
}
\description{
Search \code{paths} for EnvLogger files (EnvLogger reports, logfiles and metadata files).
}
\section{Details}{

\code{env_ls()} determines if the paths provided point to EnvLogger reports, logfiles, metadata files or none of the three.
This is done by checking for the presence of certain strings in the header of each file (such as "www.electricblue.eu, Portugal").
As a result, reports generated using the the earliest versions of the EnvLogger_Viewer app may not be recognized.
Also, if the structure of a file has been changed (e.g., by opening and saving in Excel), it may no longer be recognized.
}

\section{Non-EnvLogger data reports}{

Otherwise unsupported files that are accompanied by an associated metadata file are assumed to be valid report files created by devices outside of the EnvLogger ecosystem, and will therefore be listed as "valid reports" (but this is no guarantee that they will be successfully read by downstream functions).
}

\examples{
paths <- env_example()
env_ls(paths)
}
