% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{env_rtc_drift}
\alias{env_rtc_drift}
\title{Correct RTC drift}
\usage{
env_rtc_drift(dat, tdiff)
}
\arguments{
\item{dat}{a tibble with time (\code{t}) and temperature (\code{temp}), as the \verb{$data} column in the output from \code{\link[=read_env_all]{read_env_all()}}}

\item{tdiff}{numeric; difference in seconds between logger RTC time (obs) and smartphone time (ref) as \code{(obs - ref)}; if obs is ticking quicker than ref, \code{tdiff} is positive.}
}
\value{
The same tibble, but now with timestamps adjusted
}
\description{
Perform linear interpolation to tidy EnvLogger data so that there's one reading every hour (00 mins, 00 secs) or other specified interval; !ALWAYS! apply before binding data from different reports, as otherwise the interpolation will generate artificial data filling any gap that may be present.
}
\examples{
path  <- env_example(no_logs = TRUE)[1]
dat   <- read_env_all(path)$data[[1]]
tdiff <- read_env_all(path)$tdiff[1]
env_rtc_drift(dat, tdiff)
}
\seealso{
\code{\link[=read_env_all]{read_env_all()}}
}
