% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_id.R
\name{parse_id}
\alias{parse_id}
\alias{parse_id_cctbon}
\title{Parse structured logger ids}
\usage{
parse_id(env, div = NULL, fields = NULL, standard = NULL)

parse_id_cctbon(env)
}
\arguments{
\item{env}{The output of a call to \code{\link[=read_env]{read_env()}} (either directly, of the \verb{$report} element).}

\item{div}{A sequence of numbers defining the length of each id element. Can be provided as a number (1111123445) or as a string ("1111123445"). Numbers must be consecutive and same numbers must be contiguous. Same numbers define the position and width of the desired element in the ids.}

\item{fields}{A list providing the names and data types of each id elements. \code{fields} must have as many elements as those outlined in \code{div}. Certain downstream functions such as \code{\link[=add_info]{add_info()}} expect a column named \code{site} coding location name. Check the \code{Fields} section below for more information on the structure of \code{fields}.}

\item{standard}{In alternative to setting \code{div} and \code{fields}, provide the name of one of the supported naming schemes and the other args will be set automatically to the expected values. Current schemes supported are: 'cctbon'. Check the details sections below for more information.}
}
\value{
A tibble with EnvLogger report data, augmented with new columns resulting from the parsing of data report ids.
}
\description{
If the logger ids used follow a stable naming scheme, \code{\link[=parse_id]{parse_id()}} can be used to generate new columns that hold each naming segment. \code{\link[=parse_id_cctbon]{parse_id_cctbon()}} provides a quick shortcut for handling data that follows the CCTBON id scheme.
}
\section{standard = CCTBON}{

If the logger ids follow scheme used by the CCTBON network (\url{https://www.coastalwarming.com/cctbon}), simply use \code{\link[=parse_id_cctbon]{parse_id_cctbon()}} instead.

To conform with the CCTBON scheme, ids must:
\itemize{
\item always be 10 chr long
\item always contain the following elements in the exact order as outlined next:
\item 5 chr (#1-5) for site name (must be unique in the network)
\item 1 chr (#6) for shore level (l, m or t = low, mid or top intertidal, s = supratidal)
\item 1 chr (#7) for solar exposure (c = cold, h = hot)
\item 2 chr (#8-9) for microhabitat replicate number (01, 02, 03, ...)
\item 1 chr (#10) for logger position replicate number (a, b, c, ...)
}
}

\section{Fields}{

We can use the CCTBON scheme to better illustrate the structure of \code{fields}. In that case:
\itemize{
\item \code{div = "1111123445"}
\item \code{fields = list(site = "fct", lvl  = c("l","m","t","s"), exp  = c("c","h"), rep1 = "num", rep2 = "chr")}
}

This will result in chars #1-5 being assigned to column \code{site}, char #6 to column \code{lvl}, char #7 to column \code{exp}, and so on.
Moreover, the data types are interpreted as follows:
\itemize{
\item \code{fct} results in a factor with levels set to the unique values available
\item \code{chr} results in a character vector (as.character)
\item \code{num} results in a numeric vector (as.numeric)
\item when the type provided has a length greater than one (as is the case for \code{lvl} and \code{exp} in the CCTBON scheme), results in a factor with levels set to the values provided as type
}

As long as (a) the number of fields provided matches the number of different numbers supplied in \code{div}, (b) \code{fields} follows the structure \verb{field name = data type}, (c) data types are one of 'chr', 'num', 'fct', or c('a', 'b'), and (d) data type conversions are possible without generating \code{NA}s, any file scheme can be parsed.
}

\examples{
env <- read_env(
   env_example(c("ptzzy", "ptzzw")),
   new_interval = 60,
   show_progress = FALSE,
   show_warnings = FALSE)

# If ids follow a supported id namind standard
parse_id(env, standard = "cctbon")
parse_id_cctbon(env) # even shorter for cctbon

# If using a different scheme
parse_id(
   env,
   div = "1111123445",
   fields = list(
       site = "fct",
       lvl  = c("l","m","t","s"),
       exp  = c("c","h"),
       rep1 = "num",
       rep2 = "chr")
       )
}
\seealso{
\code{\link[=read_env]{read_env()}}, \code{\link[=add_info]{add_info()}}, \code{\link[=summarise_env]{summarise_env()}}
}
