% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_env}
\alias{plot_env}
\title{Plot EnvLogger data}
\usage{
plot_env(
  env,
  var = "temp",
  col_by = NULL,
  cols = NULL,
  tz = 0,
  same_days = FALSE,
  gg_ribbon = FALSE,
  gg_ribbon_alpha = 0.3,
  gg_facet_row = NULL,
  gg_facet_col = NULL,
  gg_cold_to_hot = TRUE,
  dy = FALSE,
  dy_focus = "all",
  dy_pointSize = NULL
)
}
\arguments{
\item{env}{The output of a call to \code{\link[=read_env]{read_env()}} (either directly, of the \verb{$report} element). Additionally, a call to \code{\link[=parse_id]{parse_id()}} or \code{\link[=add_info]{add_info()}} ensures more grouping options. In alternative, the output of a call to \code{summarise_env}.}

\item{var}{The name of the single column containing the variable to be used for plotting. Defaults to \code{temp}. Can also be set to \code{hum} or \code{press} if wanting to visualize humidity or pressure data, respectively. Can also be any other name, as long as it points to a column holding numeric data.}

\item{col_by}{The name of the single column containing the variable to be used for coloring each time series. Defaults to \code{sgmnt}, which results in the different segments of each id being shown with alternating colors (the data from each original data file corresponds to a segment, which remain identifiable even after joining by serial and id during import with \code{\link[=read_env]{read_env()}}). Coloring by segment facilitates the inspection of whether or not data has been joined together correctly). Other typical values for \code{col_by} include \code{site}, \code{lvl} and \code{exp}.}

\item{cols}{A named vector with the colors to be used. Defaults to \code{NULL}, in which case colors are determined automatically. Names must match the categories present in the column being used for coloring (define using \code{col_by}).}

\item{tz}{A number of hours to be added to timestamps (pos or neg). Please note that this doesn't provide true timezone support, as all timestamps are shifted by the same amount throughout the entire set of data.}

\item{same_days}{Should the plot only include data for days for which all distinct ids have data?}

\item{gg_ribbon}{(requires \code{dy = FALSE}) Whether to add a ribbon. If \code{TRUE}, \code{env} must include two columns strictly named \verb{$ymin} and \verb{$ymax}.}

\item{gg_ribbon_alpha}{Set the transparency of ribbons (\code{0} = fully transparent, \code{1} = fully opaque).}

\item{gg_facet_row, gg_facet_col}{(requires \code{dy = FALSE}) The name of the single column containing the variable to be used for distributing time series across different panels.}

\item{gg_cold_to_hot}{(requires \code{dy = FALSE}) Whether to color time series according to their mean \code{var} value (blue for lower values, red for higher values).}

\item{dy}{Whether to plot using \code{dygraphs}. Defaults to \code{FALSE}, resulting in a plot using \code{ggplot2}.}

\item{dy_focus}{(requires \code{dy = TRUE}) Either the names or the indexes pointing to one or more elements of the column \code{id}. Only ids referenced in \code{focus} will show in color, all other ids will be shown in black. If set to \code{all}, all ids are shown in color.}

\item{dy_pointSize}{(requires \code{dy = TRUE}) The size of points. Defaults to \code{NULL}, which cases for points not to be drawn. If provided, points are drawn to the size specified.}
}
\value{
A ggplot or a dygraph.
}
\description{
Takes an EnvLogger tibble resulting from a call to \code{\link[=read_env]{read_env()}} and produces a plot based on the \code{ggplot2} package or the \code{dygraphs} package. Use \code{dygraphs} for an interactive plot, which is most useful when the focus is on inspecting data across different time periods. Use \code{ggplot2} for more options and customization, such as adding ribbons and faceting.
}
\examples{
## DYGRAPHS
# interactivelly inspect segment joins with dygraphs
env <- read_env(
   env_example("ptzzy"),
   show_progress = FALSE,
   show_warnings = FALSE)
env <- parse_id_cctbon(env)

# note how different segments of the same id are colored differently
# (segments = data from individual EnvLogger data report files)
# zoom to 2024-01-10 to 15 to see the the transition from ...
# ... segment 1 to segment 2 for id ptzzymc01a (from green to purple)
# use dy_focus to evidence a specific id (all other ids are plotted in black)
plot_env(env, dy = TRUE, dy_focus = "ptzzymc01a", dy_pointSize = 2)

## GGPLOT2
# more flexible plots with ggplot2
env <- read_env(
   env_example(c("ptzzy", "ptzzw", "nozzz")), # 2 sites in Portugal and 1 in Norway
   new_interval = 60,
   show_progress = FALSE,
   show_warnings = FALSE)

env <- parse_id_cctbon(env)

# plot summarised data to make better sense of large datasets
# and use ribbons to depict ranges
# daily minimum, average and maximum for each site
daily_stats <- summarise_env(
   env,
   var = "temp",
   fun_list = list(
       ymin = min,
       avg = mean,
       ymax = max
       ),
   by_day = TRUE,
   by_site = TRUE
   )
plot_env(daily_stats, var = "avg", gg_ribbon = TRUE, col_by = "site")

# use same_days to restrict plot to the common data range
# provide a custom set of colors
plot_env(
   daily_stats,
   var = "avg",
   gg_ribbon = TRUE,
   col_by = "site",
   same_days = TRUE,
   cols = c(nozzz = "blue", ptzzy = "black", ptzzw = "red"))
}
\seealso{
\code{\link[=read_env]{read_env()}}, \code{\link[=parse_id]{parse_id()}}, \code{\link[=add_info]{add_info()}}, \code{\link[=summarise_env]{summarise_env()}}
}
