% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_env.R
\name{read_env}
\alias{read_env}
\title{Read EnvLogger data}
\usage{
read_env(
  paths,
  avoid_pattern = NULL,
  log_summary = FALSE,
  show_progress = TRUE,
  show_warnings = TRUE,
  read_data = TRUE,
  apply_fixes = TRUE,
  qual_lims = list(min = -30, max = +70, t0 = "2000-01-01", t1 = Sys.Date()),
  correct_rtc_drift = FALSE,
  new_interval = NULL,
  new_interval_margin = 5 * 60,
  join_serials = TRUE,
  join_ids = TRUE,
  overlap_max_mins = 60 * 3,
  keep_flags = FALSE,
  full_days = FALSE,
  auto_generate_fixes = FALSE,
  auto_implement_fixes = FALSE
)
}
\arguments{
\item{paths}{A character vector of one or more paths. May include folders, which are searched recursively.}

\item{avoid_pattern}{A character vector of one or more patterns. File paths matching to \code{avoid_pattern} are dropped.}

\item{log_summary}{Should logfile entries be provided in full (\code{FALSE}) or should they be made more compact by only including the last entry for each serial?}

\item{show_progress}{Should progress messages be printed?}

\item{show_warnings}{If issues are detected, should final warning messages be printed?}

\item{read_data}{If set to \code{FALSE}, only the header information is gathered from the data reports. This is much quicker, and can be used to assess structure and issues from a large database before importing all data.}

\item{apply_fixes}{Whether to implement the fixes found in the metadata files.}

\item{qual_lims}{A list with the elements \code{min}, \code{max}, \code{t0}, \code{t1} (names are strict). These values will be used to identify quality issues in each data report (values outside the windows provided are flagged as issues). \code{min} and \code{max} correspond to the low and high temperature thresholds. \code{t0} and \code{t1} correspond to the time window accepted for the timestamps. Default values are \code{min = -30}, \code{max = +70}, \code{t0 = "2000-01-01"} and \code{t1 = Sys.Date()}, i.e., today.}

\item{correct_rtc_drift}{Whether to expand or compress timestamps to account for recorded RTC drift.}

\item{new_interval}{Desired sampling rate in minutes. If left \code{NULL} no interpolation takes place and data remains raw. If provided, data is interpolated to every \code{new_interval}-minutes. This is important for analyses that require data from all loggers to be at the same time points (every hour, every half-hour). Interpolation is aware of eventual gaps in the data and does not generate artificial data where none exists.}

\item{join_serials}{Whether to join logger reports relating to the same serial numbers.}

\item{join_ids}{If set to \code{TRUE}, \code{join_serials} is forced to \code{TRUE} as well. Whether to join logger reports relating to the same ids (information specific to individual report files is discarded).}

\item{overlap_max_mins}{Maximum allowed overlap length in minutes. When data from two files overlaps by less than the stipulated number of minutes, the situation is normal and handled silently. Otherwise, an issue is reported.}

\item{keep_flags}{Whether to retain columns flagging data processing steps, which are normally only used internally by \code{\link[=read_env]{read_env()}} and discarded before printing the output. Only relevant for debugging errors.}

\item{full_days}{Should logger data only be kept for days in which all data points were collected? (e.g., if sampling rate is 60 minutes, full days must have 24 data points).}

\item{auto_generate_fixes}{If issues are identified for which a simple fix can be automatically determined, should the corresponding metadata files be automatically created?}

\item{auto_implement_fixes}{If set to \code{TRUE}, \code{auto_generate_fixes} is forced to \code{TRUE} as well. After automatically generating metadata files to address issues identified, should data be re-imported again to immediately implement those fixes?}
}
\value{
A list with EnvLogger information (and data) from all reports and logfiles found in the target path, divided into \verb{$rep}and \verb{$log} or, if \code{just_rep = TRUE}, a the tibble with information for all reports found.
}
\description{
Reads and organizes EnvLogger data and associated information. The following steps are executed sequentially:
\itemize{
\item \verb{step 1} -- read files
\item \verb{step 2} -- apply fixes
\item \verb{step 3} -- check quality
\item \verb{step 4} -- correct RTC drift
\item \verb{step 5} -- interpolate to round time
\item \verb{step 6} -- join by serial
\item \verb{step 7} -- join by id
\item \verb{step 8} -- keep only full days
\item \verb{step 9} -- auto-generate metadata files
}
}
\section{RTC drift correction}{

The correction implemented is linear and assumes that RTC drift accumulates at the same rate throughout the deployment. This assumption comes from the fact that all EnvLoggers feature temperature-compensated RTCs, which should result in drift being independent of temperature and only be affected by deployment length. EnvLoggers T2.4 accumulate about 10 to 15 minutes of RTC drift over a one-year deployment, which may be important to correct. EnvLoggers T7.3 and beyond feature a much more precise RTC and accumulate only 1 to 2 minutes of drift every year.
The correction applied is based on \code{tdiff}, a numeric value in seconds that represents the difference between logger RTC time (\code{obs}; observed) and smartphone time (\code{ref}; refference) as \code{(obs - ref)}. This means that if the logger RTC (\code{obs}) ticks quicker than \code{ref}, \code{tdiff} becomes positive.
\itemize{
\item drift = logger - smartphone
\item pos -> logger time was faster -> timestamps must be compressed
\item neg -> logger time was slower -> timestamps must be expanded
}
}

\section{Interpolation to round times}{

Perform linear interpolation to tidy EnvLogger data so that there's one reading every hour (00 mins, 00 secs) or other specified interval; !ALWAYS! applied before joining data from different reports, as otherwise the interpolation would generate artificial data filling any gap that may be present.
}

\examples{
paths <- env_example("ptzzy", just_dir = TRUE)
read_env(paths, log_summary = TRUE)
}
\seealso{
\code{\link[=plot_env]{plot_env()}}
}
