% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_env_all}
\alias{read_env_all}
\title{Read an EnvLogger report or logfile}
\usage{
read_env_all(path, zero_secs = FALSE, read_data = TRUE, log_summary = FALSE)
}
\arguments{
\item{path}{path to an EnvLogger report or logfile}

\item{zero_secs}{logical, defaults to \code{FALSE}; whether to remove trailing seconds (adjustment determined based on the first timestamp).}

\item{read_data}{logical, defaults to \code{TRUE}; whether to also read the data available on EnvLogger reports or to skip that step (quicker).}

\item{log_summary}{logical, defaults to \code{FALSE}; if set to \code{TRUE}, a summary of the serials included in the logfile is returned instead.}
}
\value{
If path points to an EnvLogger report, a tibble with 1 row and several columns, among which \code{id} and \code{data}. If path points to an EnvLogger logfile, a tibble with one row for each interaction with an EnvLogger and several columns with relevant metadata.
}
\description{
Given a file path, check if it points to an EnvLogger report or logfile and, if \code{TRUE}, read it.
}
\examples{
path <- env_example(no_logs = TRUE)[1]
read_env_all(path, zero_secs = TRUE) # an EnvLogger file

path <- env_example("log_")[1]
read_env_all(path) # an EnvLogger logfile

path <- env_example("corrr_")[1]
read_env_all(path) # an RTC corrections file
}
\seealso{
\code{\link[=read_env_header]{read_env_header()}}, \code{\link[=read_env_data]{read_env_data()}}, \code{\link[=read_env_log]{read_env_log()}}, \code{\link[=plot_env]{plot_env()}}
}
