% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_env_data}
\alias{read_env_data}
\title{Read the data from an EnvLogger report}
\usage{
read_env_data(path, skip, pressure = FALSE, zero_secs = TRUE, check = TRUE)
}
\arguments{
\item{path}{path to an EnvLogger report}

\item{skip}{numeric; a value indicating how many rows to skip when reading the target EnvLogger report to skip the header (usually, the \verb{$skip} column of the output from \code{read_env_header()}).}

\item{pressure}{logical, defaults to \code{FALSE}; indicates if the data being read includes a pressure field.}

\item{zero_secs}{logical, defaults to \code{TRUE}; whether to remove trailing seconds (adjustment determined based on the first timestamp).}

\item{check}{logical, defaults to \code{TRUE}; whether or not to check if \code{path} points to an EnvLogger report. Only meant to be \code{FALSE} when called from within \code{read_env_all()}, so as not to repeat the same check multiple times. \strong{Do not set it to \code{FALSE} when running this function directly.}
#'}
}
\value{
A tibble with 2 columns (\code{t} and \code{temp}).
}
\description{
Given a file path, check if it points to an EnvLogger file and, if \code{TRUE}, read just the data
}
\examples{
path <- env_example(no_logs = TRUE)[1]
read_env_header(path)
}
\seealso{
\code{\link[=read_env_header]{read_env_header()}}, \code{\link[=read_env_all]{read_env_all()}}, \code{\link[=plot_env]{plot_env()}}
}
