% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_env_header}
\alias{read_env_header}
\title{Read the header of an EnvLogger report}
\usage{
read_env_header(path, check = TRUE)
}
\arguments{
\item{path}{path to an EnvLogger report}

\item{check}{logical, defaults to \code{TRUE}; whether or not to check if \code{path} points to an EnvLogger report. Only meant to be \code{FALSE} when called from within \code{read_env_all()}, so as not to repeat the same check multiple times. \strong{Do not set it to \code{FALSE} when running this function directly.}
#'}
}
\value{
A tibble with 1 row and 12 columns with the metadata values for most of the fields recorded in the header.
}
\description{
Given a file path, check if it points to an EnvLogger report and, if \code{TRUE}, read just the header
}
\examples{
path <- env_example(no_logs = TRUE)[1]
read_env_header(path)
}
\seealso{
\code{\link[=read_env_data]{read_env_data()}}, \code{\link[=read_env_all]{read_env_all()}}, \code{\link[=plot_env]{plot_env()}}
}
