% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_env_log}
\alias{read_env_log}
\title{Read an EnvLogger logfile}
\usage{
read_env_log(path, log_summary = FALSE, check = TRUE)
}
\arguments{
\item{path}{path to an EnvLogger logfile}

\item{log_summary}{logical, defaults to \code{FALSE}; if set to \code{TRUE}, a summary of the serials included in the logfile is returned instead.}

\item{check}{logical, defaults to \code{TRUE}; whether or not to check if \code{path} points to an EnvLogger report. Only meant to be \code{FALSE} when called from within \code{read_env_all()}, so as not to repeat the same check multiple times. \strong{Do not set it to \code{FALSE} when running this function directly.}
#'}
}
\value{
A tibble with one row for each interaction with an EnvLogger and 15 columns with relevant metadata.
}
\description{
Given a file path, check if it points to an EnvLogger logfile and, if \code{TRUE}, read it.
}
\examples{
path <- env_example("log_")[1]
read_env_log(path)
}
\seealso{
\code{\link[=read_env_header]{read_env_header()}}, \code{\link[=read_env_data]{read_env_data()}}, \code{\link[=read_env_all]{read_env_all()}}, \code{\link[=plot_env]{plot_env()}}
}
