% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_env_rtc}
\alias{read_env_rtc}
\title{Read an RTC corrections file}
\usage{
read_env_rtc(path, check = TRUE)
}
\arguments{
\item{path}{path to an user-provided RTC corrections file}

\item{check}{logical, defaults to \code{TRUE}; whether or not to check if \code{path} points to an EnvLogger report. Only meant to be \code{FALSE} when called from within \code{read_env_all()}, so as not to repeat the same check multiple times. \strong{Do not set it to \code{FALSE} when running this function directly.}
#'}
}
\value{
A tibble with RTC correction parameters.
}
\description{
Given a file path, check if it points to an user-provided, correctly formatted RTC corrections file and, if \code{TRUE}, read it.
}
\examples{
path <- env_example("corrr_")[1]
read_env_rtc(path)
}
\seealso{
\code{\link[=read_env_header]{read_env_header()}}, \code{\link[=read_env_data]{read_env_data()}}, \code{\link[=read_env_all]{read_env_all()}}
}
