% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{env_example}
\alias{env_example}
\title{Get paths to example files}
\usage{
env_example(pattern = ".", just_dir = FALSE, delete_new_metadata_files = FALSE)
}
\arguments{
\item{pattern}{Patterns to select one or more example files. \code{pattern} is vectorized, so all values supplied are used. If \code{NULL} (default) all example files are returned.}

\item{just_dir}{If \code{TRUE} only folder paths are returned.}

\item{delete_new_metadata_files}{When executing the examples provided for some functions in the \link{envlogger-package}, new metadata files may be created, which in turn may limit the scope of other examples. If set to \code{TRUE}, any metadata file that isn't part of the original example files bundle is deleted, allowing for a clean slate.}
}
\value{
A named list with the full path to one or more example files, grouped by EnvLogger file type. If \code{just_dir = TRUE}, a character vector with folder paths.
}
\description{
The \link{envlogger-package} comes bundled with several sample files in its inst/extdata directory. \code{env_example()} provides a quick shortcut to access the paths to these sample files.
}
\section{Details}{

The sample files provided consist of EnvLogger reports, logfiles and metadata files concerning several years of data from a few rocky shores in northern Portugal and Norway.
The content in some of the files has been modified to introduce issues that exemplify some of the package's quality control features.
Files include:
\itemize{
\item "normal" data, just temperature - \code{nozzz}, \code{ptzzw}, and \code{ptzzy}
\item "normal" data, just temperature, different format - \code{notenv}
\item "normal" data, temperature and pressure - \code{press}
\item "normal" data, temperature and humidity - \code{humid}
\item "abnormal" data, with several quality issues - \code{issue}
}
}

\examples{
# Get the file paths of all example files
env_example()

# Get the file paths to a target serial
env_example("04CB_CC00_1507_0C")

# Get the file paths to any example files matching a search string
env_example("20250114")

# Get only logfiles
env_example()$log

# Get only folder paths
env_example(just_dir = TRUE)

# 'env_example()' is vectorized, meaning that multiple search strings can be used at once
env_example(c("ptzzy", "nozzz"))
}
