% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envlogger-package.R
\docType{package}
\name{envlogger-package}
\alias{envlogger}
\alias{envlogger-package}
\title{envlogger: Read, process and plot EnvLogger data}
\description{
Functions to handle EnvLogger data, from reading, processing and plotting. If data conforms to the CCTBON naming scheme, additional functionalities are available.
}
\author{
\strong{Maintainer}: Rui Seabra \email{ptruimar@gmail.com} (\href{https://orcid.org/0000-0002-0240-3992}{ORCID})

Authors:
\itemize{
  \item Fernando Lima \email{fplima@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})
}

}
\keyword{internal}
