% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{is_envlogger}
\alias{is_envlogger}
\title{Check if a path points to an EnvLogger file, logfile or none of the two}
\usage{
is_envlogger(paths)
}
\arguments{
\item{paths}{character vectors representing the file paths to one or several files}
}
\value{
Numeric vector indicating the type of file encountered: \code{1} if an EnvLogger file, \code{2} if an EnvLogger logfile, and \code{0} if none of the two.
}
\description{
Determine if the paths provided point to EnvLogger files.
This is done by checking for the presence of certain strings in the header of each file (such as "www.electricblue.eu, Portugal").
As a result, reports generated using the the earliest versions of EnvLogger_Viewer will not be recognized.
Also, if the structure of a file has been changed (e.g., by opening and saving in Excel), it may no longer be recognized as a report or logfile.
}
\examples{
path <- env_example(no_logs = TRUE)[1]
is_envlogger(path) # 1

path <- env_example("log_")[1]
is_envlogger(path) # 2
}
