% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{READ_ENV}
\alias{READ_ENV}
\title{Read EnvLogger data}
\usage{
READ_ENV(
  paths,
  zero_secs = FALSE,
  recursive = TRUE,
  read_data = TRUE,
  bind_id = TRUE,
  correct_rtc_drift = TRUE,
  use_rtc_correction_file = TRUE,
  approx_hourly = TRUE,
  log_summary = TRUE,
  just_rep = FALSE
)
}
\arguments{
\item{paths}{paths to EnvLogger file(s) and/or folder(s) with EnvLogger data.}

\item{zero_secs}{logical, defaults to \code{FALSE}; whether to remove trailing seconds (adjustment determined based on the first timestamp).}

\item{recursive}{logical, defaults to \code{TRUE}; whether to search for EnvLogger files recursively or not.}

\item{read_data}{logical, defaults to \code{TRUE}; whether to also read the data available on EnvLogger reports or to skip that step (quicker).}

\item{bind_id}{logical, defaults to \code{TRUE}; wether to row_bind logger reports for the same id (information specific to individual report files is discarded)}

\item{correct_rtc_drift}{logical, defaults to \code{TRUE}; if \code{TRUE}, logger data timestamps are stretched to account for the recorded clock drift}

\item{use_rtc_correction_file}{logical, defaults to \code{TRUE}; if \code{TRUE}, whenever a properly-formatted rtc corrections file is found inside of a data folder, those corretions are read and applied}

\item{approx_hourly}{logical, defaults to \code{TRUE}; if \code{TRUE}, logger data is interpolated to rounded hours (i.e., is made perfectly hourly and trimmed at the first and last days, to ensure that included days are complete; if \code{bind_id = TRUE}, trimming is executed after binding)}

\item{log_summary}{logical, defaults to \code{FALSE}; if set to \code{TRUE}, a summary of the serials included in the logfile is returned instead.}

\item{just_rep}{logical, defaults to \code{FALSE}; if \code{FALSE}, a list of all reports and all logfile data is returned, while if \code{TRUE}, only the tibble with data from reports is returned.}
}
\value{
A list with EnvLogger information (and data) from all reports and logfiles found in the target path, divided into \verb{$rep}and \verb{$log} or, if \code{just_rep = TRUE}, a the tibble with information for all reports found.
}
\description{
Wrapper function for \code{\link[=read_env_all]{read_env_all()}}, allowing for a simpler reading of all EnvLogger information and data available in a given directory.
}
\examples{
paths <- env_example("ptzzy", dir = TRUE)
READ_ENV(paths, log_summary = TRUE)
}
\seealso{
\code{\link[=read_env_all]{read_env_all()}}, \code{\link[=read_env_header]{read_env_header()}}, \code{\link[=read_env_data]{read_env_data()}}, \code{\link[=read_env_log]{read_env_log()}}, \code{\link[=plot_env]{plot_env()}}
}
