% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_information.R
\name{summarise_env}
\alias{summarise_env}
\title{Group EnvLogger data by id elements and summarise it}
\usage{
summarise_env(
  env,
  fun_list = NULL,
  site = "site",
  group = c("lvl", "exp"),
  by_day = TRUE,
  by_site = FALSE,
  by_group = FALSE,
  roll_days = 0
)
}
\arguments{
\item{env}{The output of a call to \code{\link[=read_env]{read_env()}} (either directly, of the \verb{$report} element). Additionally, a subsequent call to \code{\link[=parse_id]{parse_id()}} ensures more grouping options.}

\item{fun_list}{Named list with one or more functions to be applied to temperatures. Each function must return a single numeric value. The names of the columns derived from applying the functions provided are taken from the names of \code{fun_list}. If \code{fun} has no names, generic names are generated.}

\item{site}{The name of the column that codes for site. If a vector with more than one element, elements are combined to form a new column by pasting values together. Only relevant if \code{by_site = TRUE}. Can be omitted if data follows the CCTBON naming scheme.}

\item{group}{One or more column names to combine as a grouping factor. If a vector with more than one element, elements are combined to form a new column by pasting values together. Only relevant if \code{by_group = TRUE}. Can be omitted if data follows the CCTBON naming scheme.}

\item{by_day}{Whether to summarise values daily or at the native sampling interval.}

\item{by_site}{Whether to group by \code{site}.}

\item{by_group}{Whether to group by \code{group}.}

\item{roll_days}{Compute a rolling mean over the x days before each data point. If \code{0} (default), no rolling mean is applied.}
}
\value{
The same structure provided as input (\code{env}), augmented with new columns resulting from the joining of the information provided.
}
\description{
Takes an EnvLogger tibble resulting from a call to \code{\link[=read_env]{read_env()}}, groups data by id elements and summarises it. In the simplest case, grouping is only done along time (\code{by_day}). However, if logger ids have been broken into id elements with \code{\link[=parse_id]{parse_id()}} or \code{\link[=cctbon_id]{cctbon_id()}}, grouping can also be done across id elements. For example, if data follows the CCTBON naming scheme, data can be grouped by \code{site}, \code{lvl} or \code{exp}. This function exists mainly to condition data for \code{\link[=plot_env]{plot_env()}}, but can also be useful in itself.
}
\examples{
env <- read_env(
   env_example(c("ptzzy", "ptzzw")),
   new_interval = 60,
   show_progress = FALSE,
   show_warnings = FALSE)

env <- cctbon_id(env)

add_info(
   env,
   info = data.frame(
       var1 = 1:2,
       var3 = 1:2,
       site = c("ptzzy", "ptzzw"),
       var2 = 1:2,
       lat = c(36, 40),
       lon = -10,
       var4 = 1:2,
       exp = 1:2)
   )
}
